completely modularized the code, removed inline html function binding ot index.js and it initialized when page loads. Also updated CSS so that buttons change color when clicked
